requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.7.0